Refactor breadcrumb interfaces/support functions to helper; add parent list to modal.
